fix(interface): Fix UBSAN on compile warning

+++ framework/src/interface/tensor/tensormap.cpp
@@ -215,10 +215,10 @@ OverlapStatus CalcOverlap(
         // Therefore, they overlap if: !(a + length_a <= b OR b + length_b <= a)
         // Which is equivalent to: a + length_a > b AND b + length_b > a
 
-        size_t pStart = pTensor->offset[dim];
-        size_t pEnd = pStart + pTensor->shape[dim] - 1;
-        size_t qStart = qTensor->offset[dim];
-        size_t qEnd = qStart + qTensor->shape[dim] - 1;
+        __int128_t pStart = pTensor->offset[dim];
+        __int128_t pEnd = pStart + pTensor->shape[dim] - 1;
+        __int128_t qStart = qTensor->offset[dim];
+        __int128_t qEnd = qStart + qTensor->shape[dim] - 1;
 
         if (pEnd < qStart || qEnd < pStart) {
             return OverlapStatus::NO_OVER_LAP;
